Исправлен int на float и вывод теперь осуществляется с 3 знаками после запятой

+++ src/exercise2.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 int main(int argc, char** argv){
 
-    int arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+    float arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
     int a=0;
     for(int j=0;j<5;++j){
         a=arr[j];
@@ -10,7 +10,7 @@ int main(int argc, char** argv){
         arr[9-j]=a;
     }
     for(int i=0;i<10;i++){
-        printf("%d ",arr[i]);
+        printf("%.3f ",arr[i]);
 
     }
     return 0;
