+++ src/exercise4.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main(int argc,char** argv){
-    int arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+    float arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
     int x = atoi(argv[1]);
     int k = atoi(argv[2]);
     for(int i=9;i!=k;i--){
@@ -10,7 +10,7 @@ int main(int argc,char** argv){
     }
     arr[k]=x;
     for(int i =0;i<10;i++){
-        printf("%d ", arr[i]);
+        printf("%.3f ", arr[i]);
     }
